requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
apscheduler==3.10.4
boto3==1.35.30
python-multipart==0.0.12
pandas==2.0.3
redis==5.1.1
python-jose==3.3.0
websockets==10.0